serve_us_fyp_github/backend_new/chatbot_new/requirements.txt
PyPDF2
sentence-transformers
numpy
llama-cpp-python